Create Lab Test 3
this was a quick test of loops and functions for a intro to C++ class; however, due to technical difficulties my allotted time to complete the test was shortened. In light of this, I used fewer functions and loops, doing the assignment in a faster and more sloppy (although just as effective) manner. 

The code is supposed to accept the # of rows of seats in an airplane, then the number of seats in each individual row, and the seats per row that are/aren't occupied. It then outputs the total number of seats, the number of seats occupied and unoccupied, and the percentage of seats occupied.